got rid of circular dependency by making simple program monolithic.
