+++ colortoypiano/colortoypiano.py
@@ -4,7 +4,7 @@ a chosen midi interface. This interface is used to print all
 incoming midi messages for that device.
 """
 
-from colortoypianoshell import ColorToyPianoShell
+from cmd import Cmd
 import mido
 from queue import Queue, Empty
 import sys
@@ -35,6 +35,37 @@ class ColorToyPiano(object):
         self.t = Thread(
             target=midi_monitor, args=(inport, self.q))
         self.t.start()
+
+
+class ColorToyPianoShell(Cmd):
+    """
+    The colortoypianoshell uses the python cmd shell and the
+    colortyopiano to interface with the user and a chosen midi interface.
+    This interface is used to print all incoming midi signals for that device.
+    """
+    intro = 'This is the color toy piano. Type help or ? to list commands.\n'
+    prompt = '(color toy piano) '
+
+    # ------- init method -------
+    def __init__(self, pi):
+        super().__init__()
+        self.piano_instance = pi
+
+    # ------- basic colortoypiano commands -------
+    def do_quit(self, arg):
+        'Gracefully shutdown the color toy piano.'
+        self.piano_instance.quit()
+        return True
+
+    # ------- open new midi port -------
+    def do_open(self, arg):
+        'Open a midi in port.'
+        self.piano_instance.open(arg)
+
+    # ------- show all in ports -------
+    def do_show(self, arg):
+        'Prints available in ports.'
+        show_in_ports()
 
 
 def midi_monitor(inport, q):
